Correcoes

+++ paralelo/main.c
@@ -158,9 +158,9 @@ int main (int argc, char **argv){
                                         if (ct->debug >= 1) printf("Server[%d] permite node gravar: %d\n", tServer, i);
                                         MPI_Ssend(&node[i].li, inteiro, MPI_INT, i, 05, MPI_COMM_WORLD);
                                         gravado = 1;
-                                        MPI_Recv(&relogio[i].tempoR, 1, MPI_DOUBLE, i, 15, MPI_COMM_WORLD, &status);
-                                        MPI_Recv(&relogio[i].tempoS, 1, MPI_DOUBLE, i, 16, MPI_COMM_WORLD, &status);
-                                        MPI_Recv(&relogio[i].tempoW, 1, MPI_DOUBLE, i, 17, MPI_COMM_WORLD, &status);
+                                        MPI_Recv(&relogio[i].tempoR, 1, MPI_FLOAT, i, 15, MPI_COMM_WORLD, &status);
+                                        MPI_Recv(&relogio[i].tempoS, 1, MPI_FLOAT, i, 16, MPI_COMM_WORLD, &status);
+                                        MPI_Recv(&relogio[i].tempoW, 1, MPI_FLOAT, i, 17, MPI_COMM_WORLD, &status);
                                         if (ct->debug >= 1) printf("Server[%d] tirando node da regiao de gravacao: %d\n", tServer, i);
                                         gravar=0;
                                     }
